Rearrangement and first pour
Modified the order of operations for terminal block to fuse to buck. Added a few things like the holes for the Roboclaws, removed the display, since we're going to wire it using a JST connector in order for it to exist somewhere else on the robot.  I added a PIC footprint that fits the PIC16F819-ISO, and generally hooked it up.

I still need to check with Dustin and make sure that the VIM4 connection to the PIC is the one we want to use.

I also need to know the exact dimensions of the PCB constraint we are designing this for.